requirements.txt
# Claude Agent Desktop — зависимости
# pip install -r requirements.txt

# GUI
customtkinter>=5.2.0

# LangChain Agent
langchain>=1.0.0
langchain-openai>=0.3.0
langchain-community>=0.3.0
langchain-core>=0.3.0
langgraph>=0.3.0

# Excel
openpyxl>=3.1.0
pandas>=2.0.0

# PDF
pymupdf>=1.24.0

# Word
python-docx>=1.1.0

# Изображения
Pillow>=10.0.0

# Поиск
duckduckgo-search>=6.0.0

# Утилиты
python-dotenv>=1.0.0

# Сборка (опционально)
# pyinstaller>=6.0       # для Windows .exe
# py2app>=0.28           # для macOS .app
